Share common logic among visual editors
